Refactored code to have a better structure for future additions
