+++ NoHitMania/Models/GridPosition.swift
@@ -1,0 +1,13 @@
+//
+//  GridPosition.swift
+//  NoHitMania
+//
+//  Created by Christian Barajas on 4/15/25.
+//
+
+import Foundation
+
+struct GridPosition {
+    var x: Int
+    var y: Int
+}
