+++ NoHitMania/Utilities/Assets.xcassets/AccentColor.colorset/Contents.json
