+++ NoHitMania/Utilities/Assets.xcassets/AppIcon.appiconset/Contents.json
