+++ NoHitMania/Utilities/Assets.xcassets/Contents.json
