+++ NoHitMania/Utilities/Constants.swift
@@ -1,0 +1,24 @@
+//
+//  Constants.swift
+//  NoHitMania
+//
+//  Created by Christian Barajas on 4/15/25.
+//
+
+import Foundation
+import UIKit
+
+struct GameConstants {
+    // Grid settings
+    static let defaultGridSize = 5
+    
+    // Player settings
+    static let playerMoveAnimationDuration = 0.2
+    
+    // Colors
+    static let gridCellColor = UIColor.systemPink
+    static let gridLineColor = UIColor.white
+    
+    // Input settings
+    static let minSwipeDistance: CGFloat = 20
+}
